TST move mock proxy implementation to a separate module
